extracted logic that sync basket with dietary filter
